requirements.txt
torch
torchvision
numpy
hydra-core
omegaconf
